pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imstr"
version = "0.1.2"
description = "Compact serialization of NumPy uint8 arrays via WebP + base64"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pillow",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["src/imstr"]
